African Countries

Database of African Countries with corresponding GDP and HDI data.

Usage

african_countries

Format

A tibble with 55 rows and 7 variables:

countries

Name of the Country

region

African Region where the Country is located

uni_count

Number of Universities in the corresponding country

best_uni_rank

Rank Africa of the best University of the country

gdp

GDP of the country in 2021 (With a few exceptions for which the data is older). Source: ourworldindata.org

gdp_per_capita

GDP per Capita of the country in 2021 (With a few exceptions for which the data is older). Source: ourworldindata.org

hdi

Human Development Index of the country in 2021. Source: ourworldindata.org